Backend/requirements.txt
fastapi==0.115.12
langchain==0.3.23
langgraph==0.3.25
langchain-google-genai==2.1.2
langchain-openai==0.3.12
langchain_postgres==0.0.13
python-dotenv==1.1.0
psycopg2-binary==2.9.10
tavily-python==0.5.4
uvicorn==0.34.0
pydantic[email]